refactor: 重构反应系统并引入月相反应处理
* 从 reaction_system.py 中移除了 ReactionSystem 类及其相关逻辑。
* 引入了新的 LunarConverter 类来处理月相反应转换。
* 更新了测试以反映新的结构，将对 ReactionSystem 的引用替换为 LunarConverter。
* 调整了伤害处理以适配月相反应，包括新增的月相伤害计算测试用例。
* 确保新的伤害流程遵循加权伤害公式和元素精通加成。
