+++ core/context.py
@@ -309,13 +309,13 @@ def create_context() -> SimulationContext:
     """
     from core.logger import SimulationLogger
     from core.registry import initialize_registry
-    from core.systems.damage_system import DamageSystem
+    from core.systems.damage import DamageSystem
     from core.systems.energy_system import EnergySystem
     from core.systems.health_system import HealthSystem
     from core.systems.manager import SystemManager
     from core.systems.moonsign_system import MoonsignSystem
     from core.systems.natlan_system import NatlanSystem
-    from core.systems.reaction_system import ReactionSystem
+    from core.systems.reaction import ReactionSystem
     from core.systems.shield_system import ShieldSystem
     from core.systems.resonance_system import ResonanceSystem
