+++ core/systems/damage/__init__.py
@@ -1,0 +1,20 @@
+"""伤害系统模块。
+
+提供伤害计算相关的类：
+- DamageContext: 伤害计算上下文
+- DamagePipeline: 常规伤害流水线
+- LunarDamagePipeline: 月曜伤害流水线
+- DamageSystem: 伤害系统入口
+"""
+
+from .context import DamageContext
+from .pipeline import DamagePipeline
+from .lunar_pipeline import LunarDamagePipeline
+from .system import DamageSystem
+
+__all__ = [
+    "DamageContext",
+    "DamagePipeline",
+    "LunarDamagePipeline",
+    "DamageSystem",
+]
