+++ core/systems/reaction/handlers/__init__.py
@@ -1,0 +1,11 @@
+"""反应处理器子模块。"""
+
+from .transformative import TransformativeHandler
+from .status import StatusHandler
+from .lunar import LunarHandler
+
+__all__ = [
+    "TransformativeHandler",
+    "StatusHandler",
+    "LunarHandler",
+]
